+++ deploy.js
@@ -1,5 +1,55 @@
 const http = require("axios")
 const exec = require('child_process').exec;
+const fs = require("fs")
+
+// 同步npmmirror的包
+async function getPackages(directoryPath) {
+    return new Promise((resolve, reject) => {
+        // 读取目录下的文件和目录列表
+        fs.readdir(directoryPath, {withFileTypes: true}, (err, files) => {
+            if (err) {
+                console.log('无法读取目录:', err);
+                reject(err)
+                return;
+            }
+
+            // 过滤仅保留目录
+            const directories = files
+                .filter(file => file.isDirectory())
+                .map(directory => directory.name);
+
+            console.log('目录列表:', directories);
+            resolve(directories)
+        });
+    })
+
+}
+
+async function getAllPackages(){
+    const base = ["fast-crud",'fast-extends']
+    const ui =await getPackages("./packages/ui")
+
+    return ui.concat(base)
+}
+
+async function sync(){
+    const packages = await getAllPackages()
+    for(const pkg of packages){
+        await http({
+            url: `http://registry-direct.npmmirror.com/@fast-crud/${pkg}/sync?sync_upstream=true`,
+            method: 'PUT',
+            headers: {
+                "Content-Type": "application/json; charset=utf-8"
+            },
+            data: {}
+        })
+        console.log(`sync success:${pkg}`)
+        await sleep(10000)
+    }
+}
+
+
+
 
 //builder
 function execute(cmd){
@@ -55,7 +105,7 @@ async function trigger(){
             data:{}
         })
         console.log(`webhook success:${webhook}`)
-        await sleep(1000)
+        await sleep(10000)
     }
 
 }
@@ -63,6 +113,9 @@ async function trigger(){
 async function  start(){
     // await build()
     console.log("等待60秒")
+    await sleep(60*1000)
+    await sync()
+    console.log("同步镜像完成，等待60秒")
     await sleep(60*1000)
     await trigger()
 }