+++ packages/fast-crud/src/components/crud/fs-row-handle.vue
@@ -266,7 +266,6 @@ export default defineComponent({
   }
 
   .n-button {
-    disply: flex;
     align-items: center;
   }
 }