+++ packages/fast-crud/src/components/toolbar/index.vue
@@ -49,7 +49,6 @@ import { useI18n } from "../../locale";
 import { Constants } from "../../utils/util.constants";
 import { ButtonProps, ButtonsProps, TableColumnsProps } from "../../d";
 import { useMerge, useUi } from "../../use";
-
 /**
  * 工具条
  */
@@ -69,13 +68,6 @@ export default defineComponent({
      **/
     buttons: {
       type: Object
-    },
-    /**
-     * 当前是否显示查询。
-     * 注意：如果要隐藏search，请配置crudOptions.search.show=false
-     */
-    search: {
-      type: Boolean
     },
     /**
      * 当前是否紧凑模式
@@ -111,7 +103,7 @@ export default defineComponent({
       type: Object as PropType<any>
     }
   },
-  emits: ["refresh", "update:search", "update:compact", "update:columns", "export"],
+  emits: ["update:columns"],
   setup(props: any, ctx) {
     const { t } = useI18n();
     const columnsFilterRef: Ref = ref();
@@ -119,48 +111,7 @@ export default defineComponent({
     const { merge } = useMerge();
     const computedButtons = computed(() => {
       const defaultButtons: ButtonsProps<void> = {
-        refresh: {
-          type: "primary",
-          icon: ui.icons.refresh,
-          title: t("fs.toolbar.refresh.title"), // '刷新',
-          circle: true,
-          click: () => {
-            ctx.emit("refresh");
-          }
-        },
-        search: {
-          type: "primary",
-          icon: ui.icons.search,
-          title: t("fs.toolbar.search.title"), // '查询显示',
-          circle: true,
-          click: () => {
-            ctx.emit("update:search", !props.search);
-          }
-        },
-        compact: {
-          type: "primary",
-          icon: ui.icons.compact,
-          title: t("fs.toolbar.compact.title"), // '紧凑模式',
-          circle: true,
-          click: () => {
-            ctx.emit("update:compact", !props.compact);
-          }
-        },
-        export: {
-          show: false,
-          type: "primary",
-          icon: ui.icons.export,
-          title: t("fs.toolbar.export.title"), // '导出',
-          circle: true,
-          click: () => {
-            ctx.emit("export");
-          }
-        },
         columns: {
-          type: "primary",
-          icon: ui.icons.columnsFilter,
-          title: t("fs.toolbar.columns.title"), // '列设置',
-          circle: true,
           click: () => {
             columnsFilterRef.value.start();
           }
@@ -168,12 +119,12 @@ export default defineComponent({
       };
 
       merge(defaultButtons, props.buttons);
-      if (defaultButtons.search) {
-        defaultButtons.search.type = props.search ? "primary" : "default";
-      }
-      if (defaultButtons.compact) {
-        defaultButtons.compact.type = props.compact ? "primary" : "default";
-      }
+      // if (defaultButtons.search) {
+      //   defaultButtons.search.type = props.search ? "primary" : "default";
+      // }
+      // if (defaultButtons.compact) {
+      //   defaultButtons.compact.type = props.compact ? "primary" : "default";
+      // }
 
       let sortArr: ButtonProps[] = [];
       for (let defaultButtonsKey in defaultButtons) {