+++ packages/fast-crud/src/use/use-expose.ts
@@ -1,5 +1,13 @@
 import { Ref, toRaw } from "vue";
-import { CrudExpose, OpenDialogProps, OpenEditContext, SetFormDataOptions } from "../d/expose";
+import {
+  CrudExpose,
+  Editable,
+  EditableAddRowOptions,
+  EditableEditColsOptions,
+  OpenDialogProps,
+  OpenEditContext,
+  SetFormDataOptions
+} from "../d/expose";
 import _ from "lodash-es";
 import logger from "../utils/util.log";
 import { useMerge } from "../use/use-merge";
@@ -12,6 +20,7 @@ import {
   PageQuery,
   PageRes,
   RemoveProps,
+  SearchOptions,
   UserPageQuery,
   UserPageRes
 } from "../d";
@@ -44,7 +53,7 @@ function useEditable(props: UseEditableProps) {
   const { ui } = useUi();
   const { t } = useI18n();
   const { merge } = useMerge();
-  const editable = {
+  const editable: Editable = {
     /**
      * 启用编辑
      * @param opts
@@ -86,10 +95,10 @@ function useEditable(props: UseEditableProps) {
     /**
      * 添加行
      */
-    addRow(opts: any) {
+    addRow(opts: EditableAddRowOptions) {
       crudExpose.getTableRef().editable.addRow(opts);
     },
-    editCol(opts: any) {
+    editCol(opts: EditableEditColsOptions) {
       crudExpose.getTableRef().editable.editCol(opts);
     },
     /**
@@ -304,6 +313,61 @@ export function useExpose(props: UseExposeProps): UseExposeRet {
       checkCrudRef();
       return crudRef.value.getSearchRef();
     },
+
+    async search(pageQuery: PageQuery, options: SearchOptions = {}) {
+      const page = pageQuery.page;
+
+      let searchFormData = pageQuery.form;
+      if (searchFormData == null) {
+        searchFormData = _.cloneDeep(crudExpose.getSearchValidatedFormData());
+        //配置searchValueResolve
+        if (crudBinding.value?.search?.columns) {
+          crudExpose.doValueResolve({ form: searchFormData }, toRaw(crudBinding.value.search.columns));
+        }
+      }
+
+      let sort = pageQuery.sort;
+      if (sort == null) {
+        sort = crudBinding.value.table.sort || {};
+      }
+
+      const query: PageQuery = { page, form: searchFormData, sort };
+      let userPageQuery: UserPageQuery = query;
+      if (crudBinding.value.request.transformQuery) {
+        userPageQuery = crudBinding.value.request.transformQuery(query);
+      }
+
+      let userPageRes: UserPageRes;
+      try {
+        if (options.silence !== true) {
+          crudBinding.value.table.loading = true;
+        }
+
+        logger.debug("pageRequest", userPageQuery);
+        userPageRes = await crudBinding.value.request.pageRequest(userPageQuery);
+      } finally {
+        if (options.silence !== true) {
+          crudBinding.value.table.loading = false;
+        }
+      }
+      if (userPageRes == null) {
+        logger.warn("pageRequest返回结果不能为空");
+        return;
+      }
+      let pageRes: PageRes = userPageRes as PageRes;
+      if (crudBinding.value.request.transformRes) {
+        pageRes = crudBinding.value.request.transformRes({
+          res: userPageRes,
+          query: userPageQuery
+        });
+      }
+
+      //valueBuild
+      if (pageRes.records) {
+        crudExpose.doValueBuilder(pageRes.records);
+      }
+      return pageRes;
+    },
     async doRefresh(props?) {
       if (crudBinding.value.request.pageRequest == null) {
         return;
@@ -320,52 +384,25 @@ export function useExpose(props: UseExposeProps): UseExposeRet {
           pageSize: crudBinding.value.pagination.pageSize
         };
       }
-
-      const searchFormData = _.cloneDeep(crudExpose.getSearchValidatedFormData());
-      //配置searchValueResolve
-      if (crudBinding.value?.search?.columns) {
-        crudExpose.doValueResolve({ form: searchFormData }, toRaw(crudBinding.value.search.columns));
-      }
-      crudExpose.doValueResolve({ form: searchFormData });
-
-      const sort = crudBinding.value.table.sort || {};
-      const query: PageQuery = { page, form: searchFormData, sort };
-      let userPageQuery: UserPageQuery = query;
-      if (crudBinding.value.request.transformQuery) {
-        userPageQuery = crudBinding.value.request.transformQuery(query);
-      }
-
-      let userPageRes: UserPageRes;
-      try {
-        crudBinding.value.table.loading = true;
-        logger.debug("pageRequest", userPageQuery);
-        userPageRes = await crudBinding.value.request.pageRequest(userPageQuery);
-      } finally {
-        crudBinding.value.table.loading = false;
-      }
-      if (userPageRes == null) {
-        logger.warn("pageRequest返回结果不能为空");
-        return;
-      }
-      let pageRes: PageRes = userPageRes as PageRes;
-      if (crudBinding.value.request.transformRes) {
-        pageRes = crudBinding.value.request.transformRes({
-          res: userPageRes,
-          query: userPageQuery
-        });
-      }
+      const pageRes = await crudExpose.search({ page });
       const { currentPage = page[ui.pagination.currentPage], pageSize = page.pageSize, total } = pageRes;
       const { records } = pageRes;
-      if (records == null || total == null) {
+      if (
+        records == null ||
+        total == null ||
+        currentPage == null ||
+        currentPage <= 0 ||
+        isNaN(currentPage) ||
+        pageSize == null ||
+        pageSize <= 0 ||
+        isNaN(pageSize)
+      ) {
         logger.error(
-          "pageRequest返回结构不正确，请配置正确的request.transformRes，期望：{currentPage, pageSize, total, records:[]},实际返回：",
+          "pageRequest返回结构不正确，请配置正确的request.transformRes，期望：{currentPage>0, pageSize>0, total, records:[]},实际返回：",
           pageRes
         );
         return;
       }
-
-      //valueBuild
-      crudExpose.doValueBuilder(records);
 
       crudBinding.value.data = records;
       if (crudBinding.value.pagination) {